+++ lib/truedat/search/filters.ex
@@ -58,6 +58,10 @@ defmodule Truedat.Search.Filters do
     {:must_not, Enum.map(values, fn {key, term_values} -> term(key, term_values) end)}
   end
 
+  defp build_filter("exists", value) do
+    {:must, %{exists: value}}
+  end
+
   defp build_filter(field, value) when field in ["updated_at", "start_date", "end_date"] do
     {:filter, Query.range(field, value)}
   end